requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
langchain>=0.0.292
langchain-community>=0.3.21
openai>=0.28.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
loguru>=0.7.2
httpx>=0.24.1
jinja2>=3.1.2
python-multipart>=0.0.6 